+++ ghlang/__init__.py
@@ -1,3 +1,3 @@
 """Visualize your GitHub language stats, blazingly fast."""
 
-__version__ = "2.5.1"
+__version__ = "2.5.2"